Initial documentation on branches in ovos-releases
Additional thoughts:

I'm not sure a dev branch is needed on ovos-releases now. The chat
discussion said: "The dev branch will always reflect the dev branch of
all submodules". However that seems to be better handled by an rc/$R
branch.

I see no need for a testing branch in ovos-releases either. At any
time a rc/$R-rc$X tag can be created which can be a test point. This
avoid bugs filed against "testing" which would be a moving target.

I'm not sure a stable "branch" really make sense either; branches
continually change. If it's stable it should be a tag. Even 'latest'
should be a tag rather than a branch.

Signed-off-by: David Greaves <david@dgreaves.com>